requirements.txt
streamlit
torch==2.7.1
torchvision==0.22.1
numpy
pillow
opencv-python
